Poseidon.Analysis.Utils

Documentation

data JackknifeMode

A datatype representing the two options for how to run the Block-Jackknife

Constructors

JackknifePerN Int

JackknifePerChromosome

type GenomPos = (Chrom, Int)

A helper type to represent a genomic position.

type GroupDef = (String, SignedEntitiesList)

type PloidyVec = Vector JannoGenotypePloidy

addGroupDefs :: [GroupDef] -> [PoseidonPackage] -> [PoseidonPackage]

parseGroupDefsFromJSON :: Object -> Parser [GroupDef]

computeAlleleCount :: GenoLine -> PloidyVec -> [String] -> [Int] -> (Int, Int)

computeAlleleFreq :: GenoLine -> PloidyVec -> [String] -> [Int] -> Maybe Double

This is the original Delete-mj Jackknife based on Busing et al. 1999. The notation is the same. theta_n is the full estimate. m_j is the weights of block j. theta_j is the estimate based on the j'th block removed.

computeJackknifeOriginal :: Double -> [Double] -> [Double] -> (Double, Double)

A simplified Jackknife formula based on an additive estimate, in which the full estimate equals the Jackknife estimate

computeJackknifeAdditive :: [Int] -> [Double] -> (Double, Double)

filterTransitions :: Monad m => Bool -> Pipe (EigenstratSnpEntry, a) (EigenstratSnpEntry, a) m r

data XerxesException

Constructors

PopConfigYamlException FilePath String

GroupDefException String

FStatException String

makePloidyVec :: JannoRows -> PoseidonIO PloidyVec