Poseidon.CLI.FStats

Documentation

data FStatSpec

A datatype to represent Summary Statistics to be computed from genotype data.

Constructors

F4Spec PopSpec PopSpec PopSpec PopSpec
F4 Statistics

F3Spec PopSpec PopSpec PopSpec
F3 Statistics

F2Spec PopSpec PopSpec
F2 Statistics

PWMspec PopSpec PopSpec
Pairwise Mismatch Rates between two groups

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

data PopSpec

A datatype to represent a group or an individual

Constructors

PopSpecGroup String
Define a group name

PopSpecInd String
Define an individual name

data FstatsOptions

A datatype representing the command line options for the F-Statistics command

Constructors

FstatsOptions

Fields

_foBaseDirs :: [FilePath]

the list of base directories to search for packages

_foJackknifeMode :: JackknifeMode

The way the Jackknife is performed

_foExcludeChroms :: [Chrom]

a list of chromosome names to exclude from the computation

_foStatSpecsDirect :: [FStatSpec]

A list of F-statistics to compute

_foStatSpecsFile :: Maybe FilePath

a file listing F-statistics to compute

_foRawOutput :: Bool

whether to output the result table in raw TSV instead of nicely formatted ASCII table/

data JackknifeMode

A datatype representing the two options for how to run the Block-Jackknife

Constructors

JackknifePerN Int
Run the Jackknife in blocks of N consecutive SNPs

JackknifePerChromosome
Run the Jackknife in blocks defined as entire chromosomes.

fStatSpecParser :: Parser FStatSpec

A parser to parse Summary Statistic specifications.

runParser :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

The most general way to run a parser over the Identity monad. runParser p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

 parseFromFile p fname
   = do{ input <- readFile fname
       ; return (runParser p () fname input)
       }

runFstats :: FstatsOptions -> IO ()

The main function running the FStats command.

readStatSpecsFromFile :: FilePath -> IO [FStatSpec]