Poseidon.CLI.List

Documentation

runList :: ListOptions -> PoseidonLogIO ()

The main function running the list command

data ListOptions

A datatype representing command line options for the list command

Constructors

ListOptions

Fields

_listRepoLocation :: RepoLocationSpec

the list of base directories to search for packages

_listListEntity :: ListEntity

what to list

_listRawOutput :: Bool

whether to output raw TSV instead of a nicely formatted table

_listIgnoreGeno :: Bool

data ListEntity

A datatype to represent the options what to list

Constructors

ListPackages

ListGroups

ListIndividuals [String]

data RepoLocationSpec

Constructors

RepoLocal [FilePath]

RepoRemote String