Poseidon.CLI.Summarise

Documentation

data SummariseOptions

A datatype representing command line options for the summarise command

Constructors

SummariseOptions

Fields

_summariseBaseDirs :: [FilePath]

_summariseRawOutput :: Bool

pacReadOpts :: PackageReadOptions

runSummarise :: SummariseOptions -> PoseidonIO ()

The main function running the janno command

summariseJannoRows :: JannoRows -> Bool -> IO ()

A function to print meaningful summary information for a list of poseidon samples

paste :: [String] -> String

A helper function to concat the first N elements of a string list in a nice way

uniqueNumber :: Ord a => [a] -> String

frequency :: Ord a => [a] -> [(a, Int)]

A helper function to determine the frequency of objects in a list (similar to the table function in R)

sortTupelsBySndDesc :: (Ord a, Ord b) => (a, b) -> (a, b) -> Ordering

printFrequency :: Show a => String -> [(a, Int)] -> String

A helper function to print the output of frequency nicely

printFrequencyMaybe :: Show a => String -> [(Maybe a, Int)] -> String

A helper function to print the output of frequency over Maybe values nicely

maybeShow :: Show a => Maybe a -> String

A helper function to unwrap a maybe

printFrequencyString :: String -> [(String, Int)] -> String

As printFrequency, but without additional quoting of strings

printFrequencyMaybeString :: String -> [(Maybe String, Int)] -> String

As printFrequencyMaybe, but without additional quoting of strings

maybeShowString :: Maybe String -> String

As maybeShow, but without additional quoting of strings