Poseidon.CLI.Xerxes.FStats

Documentation

data FStatSpec

A datatype to represent F-Statistics to be computed from genotype data.

Constructors

FStatSpec FStatType [PoseidonEntity] (Maybe AscertainmentSpec)

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

runParser :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

The most general way to run a parser over the Identity monad. runParser p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

 parseFromFile p fname
   = do{ input <- readFile fname
       ; return (runParser p () fname input)
       }

data FstatsOptions

A datatype representing the command line options for the F-Statistics command

Constructors

FstatsOptions

Fields

_foBaseDirs :: [FilePath]

the list of base directories to search for packages ^ The way the Jackknife is performed

_foJackknifeMode :: JackknifeMode

The way the Jackknife is performed ^ a list of chromosome names to exclude from the computation

_foExcludeChroms :: [Chrom]

a list of chromosome names to exclude from the computation ^ A list of F-statistics to compute

_foStatInput :: [FStatInput]

A list of F-statistics to compute, entered directly or via files

_foMaxSnps :: Maybe Int

_foNoTransitions :: Bool

_foTableOut :: Maybe FilePath

_foBlockTableOut :: Maybe FilePath

data JackknifeMode

A datatype representing the two options for how to run the Block-Jackknife

Constructors

JackknifePerN Int

JackknifePerChromosome

runFstats :: FstatsOptions -> PoseidonIO ()

The main function running the FStats command.

collectStatSpecGroups :: [FStatSpec] -> [PoseidonEntity]