Poseidon.Core.Janno

Documentation

data JannoRow

A data type to represent a sample/janno file row See https://github.com/poseidon-framework/poseidon2-schema/blob/master/janno_columns.tsv for more details

Constructors

JannoRow

Fields

jPoseidonID :: PoseidonID

jGeneticSex :: GeneticSex

jGroupName :: ListColumn GroupName

jIndividualID :: Maybe JannoIndividualID

jSpecies :: Maybe JannoSpecies

jAlternativeIDs :: Maybe (ListColumn JannoAlternativeID)

jAlternativeIDsContext :: Maybe (ListColumn JannoAlternativeIDContext)

jRelationTo :: Maybe (ListColumn JannoRelationTo)

jRelationDegree :: Maybe (ListColumn JannoRelationDegree)

jRelationType :: Maybe (ListColumn JannoRelationType)

jCollectionID :: Maybe (ListColumn JannoCollectionID)

jCustodianInstitution :: Maybe (ListColumn JannoCustodianInstitution)

jCulturalEra :: Maybe (ListColumn JannoCulturalEra)

jCulturalEraURL :: Maybe (ListColumn JannoCulturalEraURL)

jArchaeologicalCulture :: Maybe (ListColumn JannoArchaeologicalCulture)

jArchaeologicalCultureURL :: Maybe (ListColumn JannoArchaeologicalCultureURL)

jCountry :: Maybe JannoCountry

jCountryISO :: Maybe JannoCountryISO

jLocation :: Maybe JannoLocation

jSite :: Maybe JannoSite

jLatitude :: Maybe JannoLatitude

jLongitude :: Maybe JannoLongitude

jDateType :: Maybe JannoDateType

jDateC14Labnr :: Maybe (ListColumn JannoDateC14Labnr)

jDateC14UncalBP :: Maybe (ListColumn JannoDateC14UncalBP)

jDateC14UncalBPErr :: Maybe (ListColumn JannoDateC14UncalBPErr)

jDateBCADStart :: Maybe JannoDateBCADStart

jDateBCADMedian :: Maybe JannoDateBCADMedian

jDateBCADStop :: Maybe JannoDateBCADStop

jChromosomalAnomalies :: Maybe (ListColumn JannoChromosomalAnomalies)

jMTHaplogroup :: Maybe JannoMTHaplogroup

jYHaplogroup :: Maybe JannoYHaplogroup

jSourceMaterial :: Maybe (ListColumn JannoSourceMaterial)

jNrLibraries :: Maybe JannoNrLibraries

jLibraryNames :: Maybe (ListColumn JannoLibraryName)

jCaptureType :: Maybe (ListColumn JannoCaptureType)

jUDG :: Maybe JannoUDG

jLibraryBuilt :: Maybe JannoLibraryBuilt

jGenotypePloidy :: Maybe JannoGenotypePloidy

jDataPreparationPipelineURL :: Maybe JannoDataPreparationPipelineURL

jEndogenous :: Maybe JannoEndogenous

jNrSNPs :: Maybe JannoNrSNPs

jCoverageOnTargets :: Maybe JannoCoverageOnTargets

jDamage :: Maybe (ListColumn JannoDamage)

jContamination :: Maybe (ListColumn JannoContamination)

jContaminationErr :: Maybe (ListColumn JannoContaminationErr)

jContaminationMeas :: Maybe (ListColumn JannoContaminationMeas)

jGeneticSourceAccessionIDs :: Maybe (ListColumn JannoGeneticSourceAccessionID)

jPrimaryContact :: Maybe JannoPrimaryContact

jPublication :: Maybe (ListColumn JannoPublication)

jComments :: Maybe JannoComment

jKeywords :: Maybe (ListColumn JannoKeyword)

jAdditionalColumns :: CsvNamedRecord

writeJannoFile :: FilePath -> JannoRows -> IO ()

writeJannoFileWithoutEmptyCols :: FilePath -> JannoRows -> IO ()

readJannoFile :: PoseidonVersion -> [ByteString] -> FilePath -> PoseidonIO JannoRows

A function to load one janno file

createMinimalJanno :: [EigenstratIndEntry] -> JannoRows

A function to create empty janno rows for a set of individuals

createMinimalSample :: EigenstratIndEntry -> JannoRow

A function to create an empty janno row for an individual

jannoHeaderString :: [String]

mainJannoColumns :: [ByteString]

newtype JannoRows

A data type to represent a janno file

Constructors

JannoRows

Fields

getJannoRows :: [JannoRow]

jannoRows2EigenstratIndEntries :: JannoRows -> [EigenstratIndEntry]

a convenience function to construct Eigenstrat Ind entries out of jannoRows

makeHeaderWithAdditionalColumns :: [JannoRow] -> Header

parseJannoRowFromNamedRecord :: PoseidonVersion -> [ByteString] -> NamedRecord -> Parser JannoRow