Poseidon.EntitiesList

Documentation

data PoseidonEntity

A datatype to represent a package, a group or an individual

Constructors

Pac String

Group String

Ind PoseidonIndividual

data SignedEntity

Constructors

Include PoseidonEntity

Exclude PoseidonEntity

type SignedEntitiesList = [SignedEntity]

type EntitiesList = [PoseidonEntity]

class Eq a => EntitySpec a

Minimal complete definition

indInfoConformsToEntitySpec, underlyingEntity, entitySpecParser

indInfoConformsToEntitySpec :: EntitySpec a => [a] -> IndividualInfo -> SelectionLevel2

underlyingEntity :: EntitySpec a => a -> PoseidonEntity

entitySpecParser :: EntitySpec a => Parser a

readEntitiesFromFile :: EntitySpec a => FilePath -> IO [a]

readEntitiesFromString :: EntitySpec a => String -> Either PoseidonException [a]

findNonExistentEntities :: EntitySpec a => [a] -> [IndividualInfo] -> EntitiesList

indInfoFindRelevantPackageNames :: EntitySpec a => [a] -> [IndividualInfo] -> [String]

filterRelevantPackages :: EntitySpec a => [a] -> [PoseidonPackage] -> [PoseidonPackage]

entitiesListP :: EntitySpec a => Parser [a]

data EntityInput a

Constructors

EntitiesDirect [a]

EntitiesFromFile FilePath

readEntityInputs :: (MonadIO m, EntitySpec a) => [EntityInput a] -> m [a]

getIndName :: PoseidonIndividual -> String

data PoseidonIndividual

Constructors

SimpleInd String

SpecificInd IndividualInfo

resolveEntityIndices :: EntitySpec a => [a] -> [IndividualInfo] -> ([[(Int, IndividualInfo, SelectionLevel2)]], [Int])

Result: fst is a list of unresolved duplicates, snd a simple list of integers for the simple single individuals

data SelectionLevel2

Constructors

ShouldBeIncluded

ShouldBeIncludedWithHigherPriority

ShouldNotBeIncluded