Poseidon.EntityTypes

Documentation

data IndividualInfo

a minimal datatype representing an individual in a collection of packages

Constructors

IndividualInfo

Fields

indInfoName :: String

the name of the individual, corresponding to jPoseidonID in Janno

indInfoGroups :: [String]

the groups associated with the individual, corresponding to jGroupName in Janno

indInfoPac :: PacNameAndVersion

the package the individual is in.

type IndividualInfoCollection = ([IndividualInfo], [Bool])

a tuple of a collection of IndividualInfos and a list of bools indicating whether the given sample is in the latest version of packages

renderNameWithVersion :: HasNameAndVersion a => a -> String

universal rendering of package names and version

class Eq a => HasNameAndVersion a where

A class to represent a package-identifying property

Methods

getPacName

Arguments

:: a

-> String
a name property

getPacVersion

Arguments

:: a

-> Maybe Version
a version property

data PoseidonEntity

A datatype to represent a requested package, group or individual

Constructors

Pac PacNameAndVersion
all individuals in a package. A version can be specified, if not implicitly request the latest

Group String
all individuals with a given group, in all of the latest packages

Ind String
all individuals with the given name, in all of the latest packages

SpecificInd String String PacNameAndVersion
the individual specified by its name, group and package. If not versioned, then take the latest version.

data SignedEntity

a signed entity specification, denoting inclusion or exclusion of an entity

Constructors

Include PoseidonEntity

Exclude PoseidonEntity

hasVersion :: HasNameAndVersion a => a -> Bool

a convenience function

type EntitiesList = [PoseidonEntity]

type SignedEntitiesList = [SignedEntity]

data PacNameAndVersion

The minimal instance of HasNameAndVersion

Constructors

PacNameAndVersion

Fields

panavName :: String

panavVersion :: Maybe Version

makePacNameAndVersion :: HasNameAndVersion a => a -> PacNameAndVersion

a function to normalise any instance of HasNameAndVersion to the minimal concrete type PacNameAndVersion

isLatestInCollection :: (MonadThrow m, HasNameAndVersion a) => [a] -> a -> m Bool

a function to check whether a given package is the latest within a collection

class EntitySpec a

A class to generalise signed and unsigned Entity Lists. Both have the feature that they can be used to filter individuals.

Minimal complete definition

indInfoConformsToEntitySpec, underlyingEntity, entitySpecParser

resolveUniqueEntityIndices :: EntitySpec a => Bool -> [a] -> IndividualInfoCollection -> PoseidonIO [Int]

indInfoConformsToEntitySpecs :: EntitySpec a => IndividualInfo -> Bool -> [a] -> Bool

this function checks whether a given individual info is selected by a given list of entities.

underlyingEntity

Arguments

:: EntitySpec a

=> a

-> PoseidonEntity
returns the unterlying entity

entitySpecParser

Arguments

:: EntitySpec a

=> Parser a
a parser

readEntitiesFromFile :: EntitySpec a => FilePath -> IO [a]

readEntitiesFromString :: EntitySpec a => String -> Either PoseidonException [a]

determineNonExistentEntities :: EntitySpec a => [a] -> IndividualInfoCollection -> EntitiesList

this returns a list of entities which could not be found

determineRelevantPackages :: (MonadThrow m, EntitySpec a) => [a] -> IndividualInfoCollection -> m [PacNameAndVersion]

determine all packages with versions that contain individuals covered by the given entities

entitiesListP :: EntitySpec a => Parser [a]

data EntityInput a

Constructors

EntitiesDirect [a]

EntitiesFromFile FilePath

readEntityInputs :: (MonadIO m, EntitySpec a, Eq a) => [EntityInput a] -> m [a]

checkIfAllEntitiesExist :: EntitySpec a => [a] -> IndividualInfoCollection -> PoseidonIO ()

resolveEntityIndices :: EntitySpec a => Bool -> [a] -> IndividualInfoCollection -> [Int]

reportDuplicateIndividuals :: [IndividualInfo] -> [(IndividualInfo, [PoseidonEntity])]

takes a list of selected individuals, checks for duplicates and reports a list of individuals with suggested Entity specifications