Poseidon.GenotypeData

Documentation

data GenoDataSource

Constructors

PacBaseDir

Fields

getPacBaseDirs :: FilePath

GenoDirect

Fields

getGenoDirect :: GenotypeDataSpec

data GenotypeDataSpec

A datatype to specify genotype files

Constructors

GenotypeDataSpec

Fields

format :: GenotypeFormatSpec

the genotype format

genoFile :: FilePath

path to the geno file

genoFileChkSum :: Maybe String

the optional checksum for the geno file

snpFile :: FilePath

path to the snp file

snpFileChkSum :: Maybe String

the optional checksum for the Snp file

indFile :: FilePath

path to the ind file

indFileChkSum :: Maybe String

the optional checksum for the indfile

snpSet :: Maybe SNPSetSpec

the SNP set de facto listed in the genotype data

data GenotypeFormatSpec

A data type representing the options fo the genotype format

Constructors

GenotypeFormatEigenstrat

GenotypeFormatPlink

data SNPSetSpec

Constructors

SNPSet1240K

SNPSetHumanOrigins

SNPSetOther

snpSetMergeList :: [SNPSetSpec] -> Bool -> SNPSetSpec

snpSetMerge :: SNPSetSpec -> SNPSetSpec -> Bool -> SNPSetSpec

loadIndividuals

Arguments

:: FilePath
the base directory

-> GenotypeDataSpec
the Genotype spec

-> IO [EigenstratIndEntry]
the returned list of EigenstratIndEntries.

A function to return a list of all individuals in the genotype files of a package.

loadGenotypeData

Arguments

:: MonadSafe m

=> FilePath
the base path

-> GenotypeDataSpec
the genotype spec

-> m ([EigenstratIndEntry], Producer (EigenstratSnpEntry, GenoLine) m ())
a pair of the EigenstratIndEntries and a Producer over the Snp position values and the genotype line.

A function to read the genotype data of a package

joinEntries :: MonadIO m => LogEnv -> [Int] -> [String] -> [Maybe (EigenstratSnpEntry, GenoLine)] -> m (EigenstratSnpEntry, GenoLine)

getConsensusSnpEntry :: MonadIO m => LogEnv -> [EigenstratSnpEntry] -> m EigenstratSnpEntry

recodeAlleles :: EigenstratSnpEntry -> EigenstratSnpEntry -> GenoLine -> Either String GenoLine

printSNPCopyProgress :: MonadIO m => LogEnv -> UTCTime -> Pipe a a m ()

selectIndices :: [Int] -> (EigenstratSnpEntry, GenoLine) -> (EigenstratSnpEntry, GenoLine)