Poseidon.Package

Documentation

data PoseidonYamlStruct

Internal structure for YAML loading only

Constructors

PoseidonYamlStruct

Fields

_posYamlPoseidonVersion :: Version

_posYamlTitle :: String

_posYamlDescription :: Maybe String

_posYamlContributor :: [ContributorSpec]

_posYamlPackageVersion :: Maybe Version

_posYamlLastModified :: Maybe Day

_posYamlGenotypeData :: GenotypeDataSpec

_posYamlJannoFile :: Maybe FilePath

_posYamlJannoFileChkSum :: Maybe String

_posYamlBibFile :: Maybe FilePath

_posYamlBibFileChkSum :: Maybe String

_posYamlReadmeFile :: Maybe FilePath

_posYamlChangelogFile :: Maybe FilePath

data PoseidonPackage

A data type to represent a Poseidon Package

Constructors

PoseidonPackage

Fields

posPacBaseDir :: FilePath

the base directory of the YAML file

posPacPoseidonVersion :: Version

the version of the package

posPacTitle :: String

the title of the package

posPacDescription :: Maybe String

the optional description string of the package

posPacContributor :: [ContributorSpec]

the contributor(s) of the package

posPacPackageVersion :: Maybe Version

the optional version of the package

posPacLastModified :: Maybe Day

the optional date of last update

posPacGenotypeData :: GenotypeDataSpec

the paths to the genotype files

posPacJannoFile :: Maybe FilePath

the path to the janno file

posPacJanno :: [JannoRow]

the loaded janno file

posPacJannoFileChkSum :: Maybe String

the optional jannofile checksum

posPacBibFile :: Maybe FilePath

the path to the BibTeX file

posPacBib :: BibTeX

the loaded bibliography file

posPacBibFileChkSum :: Maybe String

the optional bibfile chksum

posPacReadmeFile :: Maybe FilePath

the path to the README file

posPacChangelogFile :: Maybe FilePath

the path to the CHANGELOG file

posPacDuplicate :: Int

how many packages of this name exist in the current collection

data PoseidonException

A Poseidon Exception data type with several concrete constructors

Constructors

PoseidonYamlParseException FilePath ParseException
An exception to represent YAML parsing errors

PoseidonPackageException String
An exception to represent a logical error in a package

PoseidonPackageVersionException FilePath String
An exception to represent an issue with a package version

PoseidonPackageMissingVersionException FilePath
An exception to indicate a missing poseidonVersion field

PoseidonIndSearchException String
An exception to represent an error when searching for individuals or populations

PoseidonGenotypeException String
An exception to represent errors in the genotype data

PoseidonGenotypeExceptionForward SomeException
An exception to represent errors in the genotype data forwarded from the sequence-formats library

PoseidonJannoRowException FilePath Int String
An exception to represent errors when trying to parse the .janno file

PoseidonJannoConsistencyException FilePath String
An exception to represent within-janno consistency errors

PoseidonCrossFileConsistencyException String String
An exception to represent inconsistencies across multiple files in a package

PoseidonCollectionException String
An exception to represent logical issues in a poseidon package Collection

PoseidonFileExistenceException FilePath
An exception to represent missing files

PoseidonFileChecksumException FilePath
An exception to represent failed checksum tests

PoseidonFStatsFormatException String
An exception type to represent FStat specification errors

PoseidonBibTeXException FilePath String
An exception to represent errors when trying to parse the .bib file

PoseidonPoseidonEntityParsingException String
An exception to indicate failed entity parsing

PoseidonForgeEntitiesException String
An exception to indicate issues in the forge selection

PoseidonEmptyForgeException
An exception to throw if there is nothing to be forged

PoseidonNewPackageConstructionException String
An exception to indicate an issue in newPackageTemplate

PoseidonRemoteJSONParsingException String
An exception to indicate failed remote info JSON parsing

PoseidonGenericException String
A catch-all for any other type of exception

PoseidonEmptyOutPacNameException
An exception to throw if the output package lacks a name

PoseidonUnequalBaseDirException FilePath FilePath FilePath
An exception to throw if genotype data files don't share a common base directory

data PackageReadOptions

Constructors

PackageReadOptions

Fields

_readOptVerbose :: Bool

whether to print verbose output

_readOptStopOnDuplicates :: Bool

whether to stop on duplicated individuals

_readOptIgnoreChecksums :: Bool

whether to ignore all checksums

_readOptIgnoreGeno :: Bool

whether to ignore missing genotype files, useful for developer use cases

_readOptGenoCheck :: Bool

whether to check the first 100 SNPs of the genotypes

_readOptIgnorePosVersion :: Bool

whether to ignore the Poseidon version of an input package.

filterDuplicatePackages

Arguments

:: MonadThrow m

=> [PoseidonPackage]
a list of Poseidon packages with potential duplicates.

-> m [PoseidonPackage]
a cleaned up list with duplicates removed. If there are ambiguities about which package to remove, for example because last Update fields are missing or ambiguous themselves, then a Left value with an exception is returned. If successful, a Right value with the clean up list is returned.

A helper function to detect packages with duplicate names and select the most up-to-date ones.

findAllPoseidonYmlFiles :: FilePath -> IO [FilePath]

readPoseidonPackageCollection

Arguments

:: PackageReadOptions

-> [FilePath]
A list of base directories where to search in

-> PoseidonLogIO [PoseidonPackage]
A list of returned poseidon packages.

a utility function to load all poseidon packages found recursively in multiple base directories. This also takes care of smart filtering and duplication checks. Exceptions lead to skipping packages and outputting warnings

getJointGenotypeData

Arguments

:: MonadSafe m

=> LogEnv
how messages should be logged

-> Bool
whether to generate an intersection instead of union of input sites

-> [PoseidonPackage]
A list of poseidon packages.

-> Maybe FilePath
a genotype file to select SNPs from

-> m ([EigenstratIndEntry], Producer (EigenstratSnpEntry, GenoLine) m ())
a pair of the EigenstratIndEntries and a Producer over the Snp position values and the genotype line, joined across all packages.

A function to read genotype data jointly from multiple packages

getJointJanno :: [PoseidonPackage] -> [JannoRow]

getJointIndividualInfo :: [PoseidonPackage] -> [IndividualInfo]

newMinimalPackageTemplate :: FilePath -> String -> GenotypeDataSpec -> PoseidonPackage

A function to create a minimal POSEIDON package

newPackageTemplate :: FilePath -> String -> GenotypeDataSpec -> Maybe (Either [EigenstratIndEntry] [JannoRow]) -> BibTeX -> IO PoseidonPackage

A function to create a more complete POSEIDON package This will take only the filenames of the provided files, so it assumes that the files will be copied into the directory into which the YAML file will be written

renderMismatch :: [String] -> [String] -> String

zipWithPadding :: a -> b -> [a] -> [b] -> [(a, b)]

writePoseidonPackage :: PoseidonPackage -> IO ()

defaultPackageReadOptions :: PackageReadOptions

readPoseidonPackage

Arguments

:: PackageReadOptions

-> FilePath
the file path to the yaml file

-> PoseidonLogIO PoseidonPackage
the returning package returned in the IO monad.

A function to read in a poseidon package from a YAML file. Note that this function calls the addFullPaths function to make paths absolute.

makePseudoPackageFromGenotypeData :: GenotypeDataSpec -> IO PoseidonPackage