Poseidon.SecondaryTypes

Documentation

poseidonVersionParser :: Parser Version

data ContributorSpec

A data type to represent a contributor

Constructors

ContributorSpec

Fields

contributorName :: String

the name of a contributor

contributorEmail :: String

the email address of a contributor

contributorORCID :: Maybe ORCID

the ORCID of a contributor

contributorSpecParser :: Parser [ContributorSpec]

data IndividualInfo

Constructors

IndividualInfo

Fields

indInfoName :: String

indInfoGroups :: [String]

indInfoPacName :: String

data GroupInfo

Constructors

GroupInfo

Fields

gName :: String

gPackageNames :: PacNameAndVersion

gNrIndividuals :: Int

data VersionComponent

Constructors

Major

Minor

Patch

data PackageInfo

Constructors

PackageInfo

Fields

pTitle :: String

pVersion :: Maybe Version

pPosVersion :: Version

pDescription :: Maybe String

pLastModified :: Maybe Day

pNrIndividuals :: Int

runParser :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

The most general way to run a parser over the Identity monad. runParser p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

 parseFromFile p fname
   = do{ input <- readFile fname
       ; return (runParser p () fname input)
       }

data ORCID

A data type to represent an ORCID see https://support.orcid.org/hc/en-us/articles/360006897674-Structure-of-the-ORCID-Identifier

Constructors

ORCID

Fields

_orcidNums :: [Char]

_orcidChecksum :: Char

data ServerApiReturnType

Constructors

ServerApiReturnType

Fields

_apiMessages :: [String]

_apiResponse :: Maybe ApiReturnData

data ApiReturnData

Constructors

ApiReturnPackageInfo [PackageInfo]

ApiReturnGroupInfo [GroupInfo]

ApiReturnExtIndividualInfo [ExtendedIndividualInfo]

ApiReturnJanno [(String, JannoRows)]

data ExtendedIndividualInfo

Constructors

ExtendedIndividualInfo

Fields

extIndInfoName :: String

extIndInfoGroups :: [String]

extIndInfoPacName :: String

extIndInfoVersion :: Maybe Version

extIndInfoAddCols :: [(String, Maybe String)]

processApiResponse :: String -> Bool -> PoseidonIO ApiReturnData

makeNameWithVersion :: HasNameAndVersion a => a -> String

class HasNameAndVersion a where

Methods

getPacName :: a -> String

getPacVersion :: a -> Maybe Version

newtype PacNameAndVersion

Constructors

PacNameAndVersion (String, Maybe Version)