Poseidon.SequencingSource

Documentation

data SSFUDG

A datatype to represent UDG in a ssf file

Constructors

SSFMinus

SSFHalf

SSFPlus

makeSSFUDG :: MonadFail m => String -> m SSFUDG

data SSFLibraryBuilt

A datatype to represent Library_Built in a janno file

Constructors

SSFDS

SSFSS

makeSSFLibraryBuilt :: MonadFail m => String -> m SSFLibraryBuilt

newtype SeqSourceRows

A data type to represent a seqSourceFile

Constructors

SeqSourceRows

Fields

getSeqSourceRowList :: [SeqSourceRow]

data SeqSourceRow

A data type to represent a row in the seqSourceFile See https://github.com/poseidon-framework/poseidon2-schema/blob/master/seqSourceFile_columns.tsv for more details

Constructors

SeqSourceRow

Fields

sPoseidonID :: JannoStringList

sUDG :: Maybe SSFUDG

sLibraryBuilt :: Maybe SSFLibraryBuilt

sGeneticSourceAccessionIDs :: AccessionID

sStudyAccession :: Maybe AccessionID

sRunAccession :: Maybe AccessionID

sSampleAlias :: Maybe String

sSecondarySampleAccession :: Maybe String

sFirstPublic :: Maybe String

sLastUpdated :: Maybe String

sInstrumentModel :: Maybe String

sLibraryLayout :: Maybe String

sLibrarySource :: Maybe String

sInstrumentPlatform :: Maybe String

sLibraryName :: Maybe String

sLibraryStrategy :: Maybe String

sFastqFTP :: Maybe JURI

sFastqASPERA :: Maybe JURI

sFastqBytes :: Maybe Integer

sFastqMD5 :: Maybe String

sReadCount :: Maybe Integer

sSubmittedFTP :: Maybe JURI

sAdditionalColumns :: CsvNamedRecord

seqSourceHeader :: [ByteString]

seqSourceHeaderString :: [String]

seqSourceRefHashMap :: HashMap ByteString ()

writeSeqSourceFile :: FilePath -> SeqSourceRows -> IO ()

A function to write one seqSourceFile

readSeqSourceFile :: FilePath -> PoseidonIO SeqSourceRows

A function to read one seqSourceFile

readSeqSourceFileRow :: FilePath -> (Int, ByteString) -> PoseidonIO (Either PoseidonException SeqSourceRow)

A function to read one row of a seqSourceFile

checkSeqSourceConsistency :: FilePath -> SeqSourceRows -> Either PoseidonException SeqSourceRows

checkSamplesUnique :: SeqSourceRows -> Bool

checkSeqSourceRowConsistency :: FilePath -> Int -> SeqSourceRow -> Either PoseidonException SeqSourceRow

checkMandatoryNotEmpty :: SeqSourceRow -> Bool