Poseidon.Utils

Documentation

data PoseidonException

A Poseidon Exception data type with several concrete constructors

Constructors

PoseidonYamlParseException FilePath ParseException
An exception to represent YAML parsing errors

PoseidonPackageException String
An exception to represent a logical error in a package

PoseidonPackageVersionException FilePath String
An exception to represent an issue with a package version

PoseidonPackageMissingVersionException FilePath
An exception to indicate a missing poseidonVersion field

PoseidonIndSearchException String
An exception to represent an error when searching for individuals or populations

PoseidonGenotypeException String
An exception to represent errors when trying to parse the genotype data

PoseidonJannoRowException FilePath Int String
An exception to represent errors when trying to parse the .janno file

PoseidonJannoConsistencyException FilePath String
An exception to represent within-janno consistency errors

PoseidonCrossFileConsistencyException String String
An exception to represent inconsistencies across multiple files in a package

PoseidonCollectionException String
An exception to represent logical issues in a poseidon package Collection

PoseidonFileExistenceException FilePath
An exception to represent missing files

PoseidonFileChecksumException FilePath
An exception to represent failed checksum tests

PoseidonFStatsFormatException String
An exception type to represent FStat specification errors

PoseidonBibTeXException FilePath String
An exception to represent errors when trying to parse the .bib file

PoseidonPoseidonEntityParsingException String
An exception to indicate failed entity parsing

PoseidonForgeEntitiesException String
An exception to indicate issues in the forge selection

PoseidonEmptyForgeException
An exception to throw if there is nothing to be forged

PoseidonNewPackageConstructionException String
An exception to indicate an issue in newPackageTemplate

PoseidonRemoteJSONParsingException String
An exception to indicate failed remote info JSON parsing

PoseidonGenericException String
A catch-all for any other type of exception

renderPoseidonException :: PoseidonException -> String