Poseidon.GenotypeData

Documentation

data GenoDataSource

Constructors

PacBaseDir

Fields

getPacBaseDir :: FilePath

GenoDirect

Fields

getGenoDirect :: GenotypeDataSpec

data GenotypeDataSpec

Constructors

GenotypeDataSpec

Fields

genotypeFileSpec :: GenotypeFileSpec

genotypeSnpSet :: Maybe SNPSetSpec

data GenotypeFileSpec

Constructors

GenotypeEigenstrat

Fields

_esGenoFile :: FilePath

_esGenoFileChkSum :: Maybe String

_esSnpFile :: FilePath

_esSnpFileChkSum :: Maybe String

_esIndFile :: FilePath

_esIndFileChkSum :: Maybe String

GenotypePlink

Fields

_plGenoFile :: FilePath

_plGenoFileChkSum :: Maybe String

_plSnpFile :: FilePath

_plSnpFileChkSum :: Maybe String

_plIndFile :: FilePath

_plIndFileChkSum :: Maybe String

GenotypeVCF

Fields

_vcfGenoFile :: FilePath

_vcfGenoFileChkSum :: Maybe String

data GenotypeOutFormatSpec

Constructors

GenotypeOutFormatPlink

GenotypeOutFormatEigenstrat

GenotypeOutFormatVCF

data SNPSetSpec

Constructors

SNPSet1240K

SNPSetHumanOrigins

SNPSetOther

snpSetMergeList :: [SNPSetSpec] -> Bool -> SNPSetSpec

snpSetMerge :: SNPSetSpec -> SNPSetSpec -> Bool -> SNPSetSpec

reduceGenotypeFilepaths :: MonadThrow m => GenotypeDataSpec -> m (FilePath, GenotypeDataSpec)

removes directories of all filenames and returns a tuple of the basename and a modified GenotypeDataSpec with pure filenames In case basedirectories do not match, this function will throw an exception

loadIndividuals

Arguments

:: FilePath
the base directory

-> GenotypeDataSpec
the Genotype spec

-> PoseidonIO [EigenstratIndEntry]
the returned list of EigenstratIndEntries.

A function to return a list of all individuals in the genotype files of a package.

findGroupNamesInVCFheader :: [ByteString] -> Maybe [ByteString]

findGeneticSexInVCFheader :: [ByteString] -> Maybe (Either String [Sex])

loadGenotypeData

Arguments

:: MonadSafe m

=> FilePath
the base path

-> GenotypeDataSpec
the genotype spec

-> m (Producer (EigenstratSnpEntry, GenoLine) m ())
a Producer over the Snp position values and the genotype line.

A function to read the genotype data of a package

vcf2eigenstratPipe :: MonadThrow m => Pipe VCFentry (EigenstratSnpEntry, GenoLine) m r

joinEntries :: MonadIO m => LogA -> [Int] -> [String] -> [Maybe (EigenstratSnpEntry, GenoLine)] -> m (EigenstratSnpEntry, GenoLine)

getConsensusSnpEntry :: MonadIO m => LogA -> [EigenstratSnpEntry] -> m EigenstratSnpEntry

recodeAlleles :: EigenstratSnpEntry -> EigenstratSnpEntry -> GenoLine -> Either String GenoLine

printSNPCopyProgress :: MonadIO m => LogA -> UTCTime -> Pipe a a m ()

selectIndices :: [Int] -> (EigenstratSnpEntry, GenoLine) -> (EigenstratSnpEntry, GenoLine)

writeVCF :: MonadSafe m => LogA -> [JannoRow] -> FilePath -> Consumer (EigenstratSnpEntry, GenoLine) m ()

createVCFentry :: MonadIO m => LogA -> [JannoRow] -> (EigenstratSnpEntry, GenoLine) -> m VCFentry