Poseidon.Janno

Documentation

data JannoRow

A data type to represent a sample/janno file row See https://github.com/poseidon-framework/poseidon2-schema/blob/master/janno_columns.tsv for more details

Constructors

JannoRow

Fields

jPoseidonID :: String

jGeneticSex :: JannoSex

jGroupName :: JannoStringList

jAlternativeIDs :: Maybe JannoStringList

jRelationTo :: Maybe JannoStringList

jRelationDegree :: Maybe JannoRelationDegreeList

jRelationType :: Maybe JannoStringList

jRelationNote :: Maybe String

jCollectionID :: Maybe String

jCountry :: Maybe String

jCountryISO :: Maybe JannoCountry

jLocation :: Maybe String

jSite :: Maybe String

jLatitude :: Maybe Latitude

jLongitude :: Maybe Longitude

jDateType :: Maybe JannoDateType

jDateC14Labnr :: Maybe JannoStringList

jDateC14UncalBP :: Maybe JannoIntList

jDateC14UncalBPErr :: Maybe JannoIntList

jDateBCADStart :: Maybe BCADAge

jDateBCADMedian :: Maybe BCADAge

jDateBCADStop :: Maybe BCADAge

jDateNote :: Maybe String

jMTHaplogroup :: Maybe String

jYHaplogroup :: Maybe String

jSourceTissue :: Maybe JannoStringList

jNrLibraries :: Maybe Int

jLibraryNames :: Maybe JannoStringList

jCaptureType :: Maybe (JannoList JannoCaptureType)

jUDG :: Maybe JannoUDG

jLibraryBuilt :: Maybe JannoLibraryBuilt

jGenotypePloidy :: Maybe JannoGenotypePloidy

jDataPreparationPipelineURL :: Maybe JURI

jEndogenous :: Maybe Percent

jNrSNPs :: Maybe Int

jCoverageOnTargets :: Maybe Double

jDamage :: Maybe Percent

jContamination :: Maybe JannoStringList

jContaminationErr :: Maybe JannoStringList

jContaminationMeas :: Maybe JannoStringList

jContaminationNote :: Maybe String

jGeneticSourceAccessionIDs :: Maybe (JannoList AccessionID)

jPrimaryContact :: Maybe String

jPublication :: Maybe JannoStringList

jComments :: Maybe String

jKeywords :: Maybe JannoStringList

jAdditionalColumns :: CsvNamedRecord

newtype JannoSex

A datatype for genetic sex

Constructors

JannoSex

Fields

sfSex :: Sex

newtype JannoList a

A general datatype for janno list columns

Constructors

JannoList

Fields

getJannoList :: [a]

data Sex

Constructors

Male

Female

Unknown

newtype BCADAge

A datatype for BC-AD ages

Constructors

BCADAge Int

newtype JannoCountry

A datatype for countries in ISO-alpha2 code format

Constructors

JannoCountry Country

makeJannoCountryUnsafe :: String -> JannoCountry

newtype Latitude

A datatype for Latitudes

Constructors

Latitude Double

newtype Longitude

A datatype for Longitudes

Constructors

Longitude Double

data JannoDateType

A datatype to represent Date_Type in a janno file

Constructors

C14

Contextual

Modern

data JannoCaptureType

A datatype to represent Capture_Type in a janno file

Constructors

Shotgun

A1240K

ArborComplete

ArborPrimePlus

ArborAncestralPlus

TwistAncientDNA

OtherCapture

ReferenceGenome

data JannoGenotypePloidy

A datatype to represent Genotype_Ploidy in a janno file

Constructors

Diploid

Haploid

newtype Percent

A datatype for Percent values

Constructors

Percent Double

data JannoUDG

A datatype to represent UDG in a janno file

Constructors

Minus

Half

Plus

Mixed

newtype JURI

A datatype to represent URIs in a janno file

Constructors

JURI String

data RelationDegree

Constructors

Identical

First

Second

ThirdToFifth

SixthToTenth

Unrelated

OtherDegree

data JannoLibraryBuilt

A datatype to represent Library_Built in a janno file

Constructors

DS

SS

MixedSSDS

Other

data AccessionID

A datatype to represent AccessionIDs in a janno file

Constructors

INSDCProject String

INSDCStudy String

INSDCBioSample String

INSDCSample String

INSDCExperiment String

INSDCRun String

INSDCAnalysis String

OtherID String

makeAccessionID :: MonadFail m => String -> m AccessionID

writeJannoFile :: FilePath -> JannoRows -> IO ()

readJannoFile :: FilePath -> PoseidonIO JannoRows

A function to load one janno file

createMinimalJanno :: [EigenstratIndEntry] -> JannoRows

A function to create empty janno rows for a set of individuals

jannoHeaderString :: [String]

newtype CsvNamedRecord

A datatype to collect additional, unpecified .janno file columns (a hashmap in cassava/Data.Csv)

Constructors

CsvNamedRecord NamedRecord

newtype JannoRows

A data type to represent a janno file

Constructors

JannoRows [JannoRow]

type JannoStringList = JannoList String

filterLookup :: FromField a => NamedRecord -> ByteString -> Parser a

filterLookupOptional :: FromField a => NamedRecord -> ByteString -> Parser (Maybe a)

getCsvNR :: CsvNamedRecord -> NamedRecord

encodingOptions :: EncodeOptions

decodingOptions :: DecodeOptions

explicitNA :: ByteString -> ByteString

A helper functions to replace empty bytestrings values in janno files with explicit "n/a"

removeUselessSuffix :: String -> String

getMaybeJannoList :: Maybe (JannoList a) -> [a]