Poseidon.Janno

Documentation

data JannoRow

A data type to represent a sample/janno file row See https://github.com/poseidon-framework/poseidon2-schema/blob/master/janno_columns.tsv for more details

Constructors

JannoRow

Fields

jPoseidonID :: String

jGeneticSex :: GeneticSex

jGroupName :: ListColumn GroupName

jAlternativeIDs :: Maybe (ListColumn JannoAlternativeID)

jRelationTo :: Maybe (ListColumn JannoRelationTo)

jRelationDegree :: Maybe (ListColumn JannoRelationDegree)

jRelationType :: Maybe (ListColumn JannoRelationType)

jRelationNote :: Maybe JannoRelationNote

jCollectionID :: Maybe JannoCollectionID

jCountry :: Maybe JannoCountry

jCountryISO :: Maybe JannoCountryISO

jLocation :: Maybe JannoLocation

jSite :: Maybe JannoSite

jLatitude :: Maybe JannoLatitude

jLongitude :: Maybe JannoLongitude

jDateType :: Maybe JannoDateType

jDateC14Labnr :: Maybe (ListColumn JannoDateC14Labnr)

jDateC14UncalBP :: Maybe (ListColumn JannoDateC14UncalBP)

jDateC14UncalBPErr :: Maybe (ListColumn JannoDateC14UncalBPErr)

jDateBCADStart :: Maybe JannoDateBCADStart

jDateBCADMedian :: Maybe JannoDateBCADMedian

jDateBCADStop :: Maybe JannoDateBCADStop

jDateNote :: Maybe JannoDateNote

jMTHaplogroup :: Maybe JannoMTHaplogroup

jYHaplogroup :: Maybe JannoYHaplogroup

jSourceTissue :: Maybe (ListColumn JannoSourceTissue)

jNrLibraries :: Maybe JannoNrLibraries

jLibraryNames :: Maybe (ListColumn JannoLibraryName)

jCaptureType :: Maybe (ListColumn JannoCaptureType)

jUDG :: Maybe JannoUDG

jLibraryBuilt :: Maybe JannoLibraryBuilt

jGenotypePloidy :: Maybe JannoGenotypePloidy

jDataPreparationPipelineURL :: Maybe JannoDataPreparationPipelineURL

jEndogenous :: Maybe JannoEndogenous

jNrSNPs :: Maybe JannoNrSNPs

jCoverageOnTargets :: Maybe JannoCoverageOnTargets

jDamage :: Maybe JannoDamage

jContamination :: Maybe (ListColumn JannoContamination)

jContaminationErr :: Maybe (ListColumn JannoContaminationErr)

jContaminationMeas :: Maybe (ListColumn JannoContaminationMeas)

jContaminationNote :: Maybe JannoContaminationNote

jGeneticSourceAccessionIDs :: Maybe (ListColumn JannoGeneticSourceAccessionID)

jPrimaryContact :: Maybe JannoPrimaryContact

jPublication :: Maybe (ListColumn JannoPublication)

jComments :: Maybe JannoComment

jKeywords :: Maybe (ListColumn JannoKeyword)

jAdditionalColumns :: CsvNamedRecord

newtype GeneticSex

A datatype for the Genetic_Sex .janno column

Constructors

GeneticSex

Fields

sfSex :: Sex

newtype GroupName

A datatype for the Group_Name .janno column

Constructors

GroupName Text

newtype ListColumn a

A general datatype for janno list columns

Constructors

ListColumn

Fields

getListColumn :: [a]

data Sex

Constructors

Male

Female

Unknown

newtype JannoCountryISO

A datatype for countries in ISO-alpha2 code format

Constructors

JannoCountryISO Country

newtype JannoLatitude

A datatype for the Latitude .janno column

Constructors

JannoLatitude Double

newtype JannoLongitude

A datatype for the Longitude .janno column

Constructors

JannoLongitude Double

data JannoDateType

A datatype for the Date_Type .janno column

Constructors

C14

Contextual

Modern

newtype JannoDateBCADMedian

A datatype for the Date_BC_AD_Median .janno column

Constructors

JannoDateBCADMedian Int

data JannoCaptureType

A datatype for the Capture_Type .janno column

Constructors

Shotgun

A1240K

ArborComplete

ArborPrimePlus

ArborAncestralPlus

TwistAncientDNA

OtherCapture

ReferenceGenome

data JannoGenotypePloidy

A datatype for the Genotype_Ploidy .janno column

Constructors

Diploid

Haploid

data JannoUDG

A datatype for the UDG .janno column

Constructors

Minus

Half

Plus

Mixed

data JannoRelationDegree

A datatype for the Relation_Degree .janno column

Constructors

Identical

First

Second

ThirdToFifth

SixthToTenth

Unrelated

OtherDegree

data JannoLibraryBuilt

A datatype for the Library_Built .janno column

Constructors

DS

SS

MixedSSDS

Other

writeJannoFile :: FilePath -> JannoRows -> IO ()

writeJannoFileWithoutEmptyCols :: FilePath -> JannoRows -> IO ()

readJannoFile :: FilePath -> PoseidonIO JannoRows

A function to load one janno file

createMinimalJanno :: [EigenstratIndEntry] -> JannoRows

A function to create empty janno rows for a set of individuals

createMinimalSample :: EigenstratIndEntry -> JannoRow

A function to create an empty janno row for an individual

jannoHeaderString :: [String]

newtype CsvNamedRecord

A datatype to collect additional, unpecified .janno file columns (a hashmap in cassava/Data.Csv)

Constructors

CsvNamedRecord NamedRecord

newtype JannoRows

A data type to represent a janno file

Constructors

JannoRows

Fields

getJannoRows :: [JannoRow]

type JannoStringList = ListColumn String

filterLookup :: FromField a => NamedRecord -> ByteString -> Parser a

filterLookupOptional :: FromField a => NamedRecord -> ByteString -> Parser (Maybe a)

getCsvNR :: CsvNamedRecord -> NamedRecord

encodingOptions :: EncodeOptions

decodingOptions :: DecodeOptions

explicitNA :: NamedRecord -> NamedRecord

removeUselessSuffix :: String -> String

parseCsvParseError :: Parser CsvParseError

renderCsvParseError :: CsvParseError -> String

getMaybeListColumn :: Maybe (ListColumn a) -> [a]

jannoRows2EigenstratIndEntries :: JannoRows -> [EigenstratIndEntry]

a convenience function to construct Eigenstrat Ind entries out of jannoRows

makeHeaderWithAdditionalColumns :: [JannoRow] -> Header