Poseidon.SecondaryTypes

Documentation

poseidonVersionParser :: Parser Version

data ContributorSpec

A data type to represent a contributor

Constructors

ContributorSpec

Fields

contributorName :: String

the name of a contributor ^ the email address of a contributor

contributorEmail :: String

the email address of a contributor

contributorSpecParser :: Parser [ContributorSpec]

data IndividualInfo

Constructors

IndividualInfo

Fields

indInfoName :: String

indInfoGroups :: [String]

indInfoPacName :: String

data GroupInfo

Constructors

GroupInfo

Fields

gName :: String

gPackageNames :: [String]

gNrIndividuals :: Int

data VersionComponent

Constructors

Major

Minor

Patch

data PackageInfo

Minimal package representation on Poseidon servers

Constructors

PackageInfo

Fields

pTitle :: String

pVersion :: Maybe Version

pDescription :: Maybe String

pLastModified :: Maybe Day

pNrIndividuals :: Int

runParser :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

The most general way to run a parser over the Identity monad. runParser p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

 parseFromFile p fname
   = do{ input <- readFile fname
       ; return (runParser p () fname input)
       }