Poseidon.SequencingSource

Documentation

newtype SeqSourceRows

A data type to represent a seqSourceFile

Constructors

SeqSourceRows

Fields

getSeqSourceRowList :: [SeqSourceRow]

data SeqSourceRow

A data type to represent a row in the seqSourceFile See https://github.com/poseidon-framework/poseidon2-schema/blob/master/seqSourceFile_columns.tsv for more details

Constructors

SeqSourceRow

Fields

sPoseidonID :: Maybe (ListColumn String)

sUDG :: Maybe SSFUDG

sLibraryBuilt :: Maybe SSFLibraryBuilt

sSampleAccession :: Maybe SSFAccessionIDSample

sStudyAccession :: Maybe SSFAccessionIDStudy

sRunAccession :: Maybe SSFAccessionIDRun

sSampleAlias :: Maybe SSFSampleAlias

sSecondarySampleAccession :: Maybe SSFSecondarySampleAccession

sFirstPublic :: Maybe SSFFirstPublicSimpleDate

sLastUpdated :: Maybe SSFLastUpdatedSimpleDate

sInstrumentModel :: Maybe SSFInstrumentModel

sLibraryLayout :: Maybe SSFLibraryLayout

sLibrarySource :: Maybe SSFLibrarySource

sInstrumentPlatform :: Maybe SSFInstrumentPlatform

sLibraryName :: Maybe SSFLibraryName

sLibraryStrategy :: Maybe SSFLibraryStrategy

sFastqFTP :: Maybe (ListColumn SSFFastqFTPURI)

sFastqASPERA :: Maybe (ListColumn SSFFastqASPERAURI)

sFastqBytes :: Maybe (ListColumn SSFFastqBytes)

sFastqMD5 :: Maybe (ListColumn SSFFastqMD5)

sReadCount :: Maybe SSFReadCount

sSubmittedFTP :: Maybe (ListColumn SSFSubmittedFTPURI)

sAdditionalColumns :: CsvNamedRecord

seqSourceHeader :: [ByteString]

seqSourceHeaderString :: [String]

seqSourceRefHashMap :: HashMap ByteString ()

writeSeqSourceFile :: FilePath -> SeqSourceRows -> IO ()

A function to write one seqSourceFile

readSeqSourceFile :: FilePath -> PoseidonIO SeqSourceRows

A function to read one seqSourceFile

readSeqSourceFileRow :: FilePath -> (Int, ByteString) -> PoseidonIO (Either PoseidonException SeqSourceRow)

A function to read one row of a seqSourceFile

warnSeqSourceConsistency :: FilePath -> SeqSourceRows -> PoseidonIO ()

checkRunsUnique :: SeqSourceRows -> Bool

checkAtLeastOnePoseidonID :: SeqSourceRows -> Bool