Poseidon.SequencingSource

Documentation

data AccessionID

A datatype to represent AccessionIDs in a ssf file

Constructors

INSDCProject String

INSDCStudy String

INSDCBioSample String

INSDCSample String

INSDCExperiment String

INSDCRun String

INSDCAnalysis String

OtherID String

makeAccessionID :: MonadFail m => String -> m AccessionID

newtype JURI

A datatype to represent URIs in a ssf file

Constructors

JURI String

makeJURI :: MonadFail m => String -> m JURI

data SSFUDG

A datatype to represent UDG in a ssf file

Constructors

SSFMinus

SSFHalf

SSFPlus

makeSSFUDG :: MonadFail m => String -> m SSFUDG

data SSFLibraryBuilt

A datatype to represent Library_Built in a janno file

Constructors

SSFDS

SSFSS

makeSSFLibraryBuilt :: MonadFail m => String -> m SSFLibraryBuilt

newtype SeqSourceRows

A data type to represent a seqSourceFile

Constructors

SeqSourceRows

Fields

getSeqSourceRowList :: [SeqSourceRow]

newtype AccessionIDRun

Constructors

AccessionIDRun

Fields

getRunAccession :: AccessionID

makeAccessionIDRun :: MonadFail m => String -> m AccessionIDRun

newtype AccessionIDSample

Constructors

AccessionIDSample

Fields

getSampleAccession :: AccessionID

makeAccessionIDSample :: MonadFail m => String -> m AccessionIDSample

newtype AccessionIDStudy

Constructors

AccessionIDStudy

Fields

getStudyAccession :: AccessionID

makeAccessionIDStudy :: MonadFail m => String -> m AccessionIDStudy

newtype SimpleDate

A datatype for calendar dates

Constructors

SimpleDate Day

makeSimpleDate :: MonadFail m => String -> m SimpleDate

newtype MD5

A datatype to represent MD5 hashes

Constructors

MD5 String

makeMD5 :: MonadFail m => String -> m MD5

isMD5Hash :: String -> Bool

data SeqSourceRow

A data type to represent a row in the seqSourceFile See https://github.com/poseidon-framework/poseidon2-schema/blob/master/seqSourceFile_columns.tsv for more details

Constructors

SeqSourceRow

Fields

sPoseidonID :: Maybe JannoStringList

sUDG :: Maybe SSFUDG

sLibraryBuilt :: Maybe SSFLibraryBuilt

sSampleAccession :: Maybe AccessionIDSample

sStudyAccession :: Maybe AccessionIDStudy

sRunAccession :: Maybe AccessionIDRun

sSampleAlias :: Maybe String

sSecondarySampleAccession :: Maybe String

sFirstPublic :: Maybe SimpleDate

sLastUpdated :: Maybe SimpleDate

sInstrumentModel :: Maybe String

sLibraryLayout :: Maybe String

sLibrarySource :: Maybe String

sInstrumentPlatform :: Maybe String

sLibraryName :: Maybe String

sLibraryStrategy :: Maybe String

sFastqFTP :: Maybe (ListColumn JURI)

sFastqASPERA :: Maybe (ListColumn JURI)

sFastqBytes :: Maybe (ListColumn Integer)

sFastqMD5 :: Maybe (ListColumn MD5)

sReadCount :: Maybe Integer

sSubmittedFTP :: Maybe (ListColumn JURI)

sAdditionalColumns :: CsvNamedRecord

seqSourceHeader :: [ByteString]

seqSourceHeaderString :: [String]

seqSourceRefHashMap :: HashMap ByteString ()

writeSeqSourceFile :: FilePath -> SeqSourceRows -> IO ()

A function to write one seqSourceFile

readSeqSourceFile :: FilePath -> PoseidonIO SeqSourceRows

A function to read one seqSourceFile

readSeqSourceFileRow :: FilePath -> (Int, ByteString) -> PoseidonIO (Either PoseidonException SeqSourceRow)

A function to read one row of a seqSourceFile

warnSeqSourceConsistency :: FilePath -> SeqSourceRows -> PoseidonIO ()

checkRunsUnique :: SeqSourceRows -> Bool

checkAtLeastOnePoseidonID :: SeqSourceRows -> Bool