Poseidon.Utils

Documentation

data PoseidonException

A Poseidon Exception data type with several concrete constructors

Constructors

PoseidonYamlParseException FilePath ParseException
An exception to represent YAML parsing errors

PoseidonPackageException String
An exception to represent a logical error in a package

PoseidonPackageVersionException FilePath String
An exception to represent an issue with a package version

PoseidonPackageMissingVersionException FilePath
An exception to indicate a missing poseidonVersion field

PoseidonIndSearchException String
An exception to represent an error when searching for individuals or populations

PoseidonGenotypeException String
An exception to represent errors in the genotype data

PoseidonGenotypeExceptionForward SomeException
An exception to represent errors in the genotype data forwarded from the sequence-formats library

PoseidonFileRowException FilePath Int String
An exception to represent errors when trying to parse the janno or seqSource file

PoseidonFileConsistencyException FilePath String
An exception to represent consistency errors in janno or seqSource files

PoseidonCrossFileConsistencyException String String
An exception to represent inconsistencies across multiple files in a package

PoseidonCollectionException String
An exception to represent logical issues in a poseidon package Collection

PoseidonFileExistenceException FilePath
An exception to represent missing files

PoseidonFileChecksumException FilePath
An exception to represent failed checksum tests

PoseidonFStatsFormatException String
An exception type to represent FStat specification errors

PoseidonBibTeXException FilePath String
An exception to represent errors when trying to parse the .bib file

PoseidonPoseidonEntityParsingException String
An exception to indicate failed entity parsing

PoseidonForgeEntitiesException String
An exception to indicate issues in the forge selection

PoseidonEmptyForgeException
An exception to throw if there is nothing to be forged

PoseidonNewPackageConstructionException String
An exception to indicate an issue in newPackageTemplate

PoseidonRemoteJSONParsingException String
An exception to indicate failed remote info JSON parsing

PoseidonGenericException String
A catch-all for any other type of exception

PoseidonEmptyOutPacNameException
An exception to throw if the output package lacks a name

PoseidonUnequalBaseDirException FilePath FilePath FilePath
An exception to throw if genotype data files don't share a common base directory

renderPoseidonException :: PoseidonException -> String

usePoseidonLogger :: LogMode -> PlinkPopNameMode -> PoseidonIO a -> IO a

testLog :: PoseidonIO a -> IO a

type PoseidonIO = ReaderT Env IO

envLogAction :: PoseidonIO LogA

envInputPlinkMode :: PoseidonIO PlinkPopNameMode

data LogMode

Constructors

NoLog

SimpleLog

DefaultLog

ServerLog

VerboseLog

checkFile :: FilePath -> Maybe String -> IO ()

getChecksum :: FilePath -> IO String

logWarning :: String -> PoseidonIO ()

logInfo :: String -> PoseidonIO ()

logDebug :: String -> PoseidonIO ()

logError :: String -> PoseidonIO ()

type LogA = LogAction IO Message

noLog :: LogA

logWithEnv :: MonadIO m => LogA -> PoseidonIO () -> m ()

padRight :: Int -> String -> String

padLeft :: Int -> String -> String

determinePackageOutName :: Maybe String -> FilePath -> IO String

data PlinkPopNameMode

Constructors

PlinkPopNameAsFamily

PlinkPopNameAsPhenotype

PlinkPopNameAsBoth